requirements.txt
haversine==2.7.0
plotly==5.10.0
pandas==1.4.3
streamlit-folium==0.7.0
folium==0.13.0
numpy==1.23.1
Pillow==9.2.0
matplotlib==3.5.3
matplotlib-inline==0.1.6
streamlit==1.15.1